# src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Agendamiento de Servicios",
    "version": "0.1.0",
    "description": "API para gestionar calendarios y eventos usando Google Calendar. Permite crear, actualizar, eliminar y consultar eventos, así como verificar disponibilidad de horarios."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor de desarrollo"
    }
  ],
  "tags": [
    {
      "name": "Sistema",
      "description": "Endpoints relacionados con el estado del sistema"
    },
    {
      "name": "Calendarios",
      "description": "Gestión de calendarios"
    },
    {
      "name": "Eventos",
      "description": "Gestión de eventos del calendario"
    },
    {
      "name": "Disponibilidad",
      "description": "Verificación de disponibilidad de horarios"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": ["Sistema"],
        "summary": "Health Check",
        "description": "Verifica el estado del servicio",
        "responses": {
          "200": {
            "description": "Servicio funcionando correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "service": {
                      "type": "string",
                      "example": "hono-google-calendar"
                    },
                    "version": {
                      "type": "string",
                      "example": "0.1.0"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2024-10-30T10:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/calendar/list": {
      "get": {
        "tags": ["Calendarios"],
        "summary": "Listar calendarios",
        "description": "Obtiene la lista de calendarios disponibles usando la cuenta de servicio de Google",
        "responses": {
          "200": {
            "description": "Lista de calendarios obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener la lista de calendarios",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/calendar/events": {
      "get": {
        "tags": ["Eventos"],
        "summary": "Obtener eventos",
        "description": "Obtiene los eventos de un calendario en un rango de tiempo específico",
        "parameters": [
          {
            "name": "calendarId",
            "in": "query",
            "description": "ID del calendario",
            "schema": {
              "type": "string",
              "example": "primary"
            }
          },
          {
            "name": "timeMin",
            "in": "query",
            "description": "Fecha y hora mínima (también se acepta periodStart)",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2024-10-30T00:00:00Z"
            }
          },
          {
            "name": "timeMax",
            "in": "query",
            "description": "Fecha y hora máxima (también se acepta periodEnd)",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2024-10-31T23:59:59Z"
            }
          },
          {
            "name": "periodStart",
            "in": "query",
            "description": "Alternativa a timeMin",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2024-10-30T00:00:00Z"
            }
          },
          {
            "name": "periodEnd",
            "in": "query",
            "description": "Alternativa a timeMax",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2024-10-31T23:59:59Z"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Eventos obtenidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "events": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener eventos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/calendar/event/{eventId}": {
      "get": {
        "tags": ["Eventos"],
        "summary": "Obtener evento específico",
        "description": "Obtiene los detalles de un evento específico por su ID",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "description": "ID del evento",
            "schema": {
              "type": "string",
              "example": "abc123xyz"
            }
          },
          {
            "name": "calendarId",
            "in": "query",
            "description": "ID del calendario",
            "schema": {
              "type": "string",
              "example": "primary"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evento obtenido exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "event": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener el evento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Eventos"],
        "summary": "Actualizar evento",
        "description": "Actualiza un evento existente en el calendario",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "description": "ID del evento a actualizar",
            "schema": {
              "type": "string",
              "example": "abc123xyz"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEvent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evento actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "event": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al actualizar el evento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Eventos"],
        "summary": "Eliminar evento",
        "description": "Elimina un evento del calendario",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "description": "ID del evento a eliminar",
            "schema": {
              "type": "string",
              "example": "abc123xyz"
            }
          },
          {
            "name": "calendarId",
            "in": "query",
            "description": "ID del calendario",
            "schema": {
              "type": "string",
              "example": "primary"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evento eliminado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Event deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al eliminar el evento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/calendar/event": {
      "post": {
        "tags": ["Eventos"],
        "summary": "Crear evento",
        "description": "Crea un nuevo evento en el calendario. Utiliza un sistema de bloqueo con Redis para evitar conflictos de reserva en el mismo horario.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evento creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "available": {
                      "type": "boolean",
                      "example": true
                    },
                    "event": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "El horario no está disponible (conflicto)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "available": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Slot busy"
                    },
                    "conflictingEvents": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al crear el evento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/calendar/check-availability": {
      "post": {
        "tags": ["Disponibilidad"],
        "summary": "Verificar disponibilidad",
        "description": "Verifica si un horario específico está disponible en el calendario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckAvailability"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Disponibilidad verificada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "available": {
                      "type": "boolean",
                      "example": true
                    },
                    "conflictingEvents": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error al verificar disponibilidad",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateEvent": {
        "type": "object",
        "required": ["startDateTime"],
        "properties": {
          "calendarId": {
            "type": "string",
            "description": "ID del calendario. Si no se proporciona, se usa el calendario por defecto",
            "example": "primary"
          },
          "startDateTime": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha y hora de inicio del evento en formato ISO 8601",
            "example": "2024-10-30T10:00:00Z"
          },
          "endDateTime": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha y hora de fin del evento. Si no se proporciona, se calcula automáticamente",
            "example": "2024-10-30T11:00:00Z"
          },
          "summary": {
            "type": "string",
            "description": "Título o resumen del evento",
            "example": "Cita médica"
          },
          "description": {
            "type": "string",
            "description": "Descripción detallada del evento",
            "example": "Consulta general con el Dr. García"
          },
          "location": {
            "type": "string",
            "description": "Ubicación del evento",
            "example": "Consultorio 101, Calle Principal #123"
          },
          "attendees": {
            "type": "array",
            "description": "Lista de asistentes al evento",
            "items": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "format": "email",
                  "example": "paciente@example.com"
                }
              }
            }
          }
        }
      },
      "UpdateEvent": {
        "type": "object",
        "properties": {
          "calendarId": {
            "type": "string",
            "description": "ID del calendario",
            "example": "primary"
          },
          "summary": {
            "type": "string",
            "description": "Nuevo título del evento",
            "example": "Cita médica actualizada"
          },
          "description": {
            "type": "string",
            "description": "Nueva descripción del evento",
            "example": "Consulta de seguimiento"
          },
          "location": {
            "type": "string",
            "description": "Nueva ubicación del evento",
            "example": "Consultorio 102"
          },
          "startDateTime": {
            "type": "string",
            "format": "date-time",
            "description": "Nueva fecha y hora de inicio",
            "example": "2024-10-30T11:00:00Z"
          },
          "endDateTime": {
            "type": "string",
            "format": "date-time",
            "description": "Nueva fecha y hora de fin",
            "example": "2024-10-30T12:00:00Z"
          },
          "attendees": {
            "type": "array",
            "description": "Nueva lista de asistentes",
            "items": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "format": "email",
                  "example": "paciente@example.com"
                }
              }
            }
          }
        }
      },
      "CheckAvailability": {
        "type": "object",
        "required": ["startDateTime"],
        "properties": {
          "calendarId": {
            "type": "string",
            "description": "ID del calendario a verificar",
            "example": "primary"
          },
          "startDateTime": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha y hora de inicio del período a verificar",
            "example": "2024-10-30T10:00:00Z"
          },
          "endDateTime": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha y hora de fin del período a verificar",
            "example": "2024-10-30T11:00:00Z"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "error_code"
          },
          "details": {
            "type": "string",
            "example": "Descripción detallada del error"
          }
        }
      }
    }
  }
}